Move key generation concerns to sub-module
Why you made the change:

I did this for one to better organized the key generation concern. However, most
importantly to facilitate sharing of the key generation logic to facilitate
creation of other types of repositories. This is especially application to what
will most likely be the new background refresh repository.
